+++ www/modules/email/MessagePanel.js
@@ -55,7 +55,6 @@ GO.email.MessagePanel = Ext.extend(Ext.Panel, {
 			'</tr>'+
 			'<tr><td><b>'+t("Subject", "email")+'</b></td><td>: {subject}</td></tr>'+
 			'<tr><td><b>'+t("Date")+'</b></td><td>: {date}</td></tr>'+
-			//'<tr><td><b>'+t("Size")+'</b></td><td>: {size}</td></tr>'+
 			'<tr><td><b>'+t("To", "email")+'</b></td><td>: '+
 			'<tpl for="to">'+
 			'{personal} <tpl if="email.length">&lt;<a href="mailto:&quot;{[GO.util.html_entity_decode(values.personal, \'ENT_QUOTES\')]}&quot; &lt;{email}&gt;">{email}</a>&gt;; </tpl>'+
@@ -94,9 +93,6 @@ GO.email.MessagePanel = Ext.extend(Ext.Panel, {
 				'<i class="icon ic-more-vert" id="downloadAllMenu-'+this.downloadAllMenuId +'"></i>'+
 			'</tpl>'+
 							
-							
-			
-			
 			'</td></tr>'+
 			'</table>'+
 			'</tpl>'+			
@@ -106,13 +102,16 @@ GO.email.MessagePanel = Ext.extend(Ext.Panel, {
 				'<h5 class="em-links-header">'+t("Links")+'</h5>'+
 				'<div class="em-links">'+
 				'<tpl for="links">'+
+					'<div class="go-icon-list"><p><i class="label entity {[this.linkIconCls(values)]}"></i> ' +
 					'<tpl if="entity==\'LinkedEmail\'">'+
-						'<div class="go-icon-list"><p><i class="label entity {[this.linkIconCls(values)]}"></i> <a href="#" onclick="const win = new go.links.LinkDetailWindow({entity\:\'LinkedEmail\'});win.load({model_id});">{name}</a> <label>{description}</label></p></div>'+
+					'<a href="#email"  onclick="const win = new go.links.LinkDetailWindow({entity\:\'LinkedEmail\'});win.load({model_id});">'+
 					'</tpl>' +
 					'<tpl if="entity!=\'LinkedEmail\'">' +
-						'<div class="go-icon-list"><p><i class="label entity {[this.linkIconCls(values)]}"></i> <a href="#{entity}/{model_id}">{name}</a> <label>{description}</label>' +
-						'</p></div>'+
+						'<a href="#{entity}/{model_id}">' +
 					'</tpl>' +
+					'{name}</a> <label>{description}</label>' +
+					'{[this.addDeleteBtn(values)]}</p>' +
+					'</div>' +
 				'</tpl>'+
 			'</div></tpl>'+
 			
@@ -220,22 +219,18 @@ GO.email.MessagePanel = Ext.extend(Ext.Panel, {
 
 			},
 			addDeleteBtn: function(link) {
-				let btn = new Ext.Button({
-					itemId: "delete",
-					iconCls: "ic-delete",
-					text: t("Unlink"),
-					handler: function () {
-						Ext.MessageBox.confirm(t("Delete"), t("Are you sure you want to unlink this item?"), function (btn) {
-							if (btn == "yes") {
-								go.Db.store("Link").set({
-									destroy: [link.id]
-								});
-							}
-						}, this);
-					},
-					scope: this
-				});
-				return btn.render();
+				return "";
+				// return '<a class="simple-link" onclick="go.Db.store(\'Link\').set({destroy: ['+link.link_id+']}).then(() => {GO.mainLayout.openModule(\'email\');});">'+t('Delete') + '</a>';
+				// TODO: Handler:
+				// 	handler: function () {
+				// 		Ext.MessageBox.confirm(t("Delete"), t("Are you sure you want to unlink this item?"), function (btn) {
+				// 			if (btn == "yes") {
+				// 				go.Db.store("Link").set({
+				// 					destroy: [link.id]
+				// 				});
+				// 			}
+				// 		}, this);
+				// 	},
 			},
 			addSlashes : function(str)
 			{