+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+07-11-2022 6.6.125
 - Core: Switch user will not update last login and login count.
 - Billing: fixed division by zero
 - Email: 16px on mobile in htmleditor so iphone won't zoom in
+- Core: fix search user and contact search in email address bar
+- Core / Address book: Always create address book entry for user
+- Core: Error dialog could display nothing when horizontal scrollbar was present
+- Core: Accent color for combo group headers
 
 31-10-2022 6.6.124
 - Core: Add extra CSS options to custom comboboxes